Former-commit-id: f56ccb35560c1a90764d49bf5365d8f07208a081 [formerly 5d59ebebf6b3b036f53165123614c54a6ab94f9f]
Former-commit-id: 726bad125158deaed82f0965a570b3426b3d60b2

+++ backend/RunWithMe/src/main/java/kr/co/rwm/controller/RunningController.java
@@ -65,13 +65,20 @@ public class RunningController {
 	
 	// redis에 userid를 key로 km당 기록을 저장한다.
 	@PostMapping("/temp")
-	public ResponseEntity<String> saveRecord(@RequestBody Record record){
+	public ResponseEntity saveRecord(@RequestBody Record record, HttpServletRequest request){
 		System.out.println("running/controller/temp/record");
-		User loginUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		record.setUserId(loginUser);
-		recordTempRepository.setUserRecord(record);
+		String token = request.getHeader("AUTH");
+		if(jwtTokenProvider.validateToken(token)) {
+			User loginUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+			record.setUserId(loginUser);
+			recordTempRepository.setUserRecord(record);
+			
+			return new ResponseEntity<String>(ResponseMessage.RECORD_REDIS_INSERT_SUCCESS, HttpStatus.CREATED);
 		
-		return new ResponseEntity<String>(ResponseMessage.RECORD_REDIS_INSERT_SUCCESS, HttpStatus.CREATED);
+		}else {
+			return new ResponseEntity<Response>(new Response(StatusCode.FORBIDDEN, ResponseMessage.UNAUTHORIZED),
+					HttpStatus.FORBIDDEN);
+		}
 	}
 	
 	   // swipe했을 때 redis에 있던 record를 보내준다.
@@ -94,44 +101,51 @@ public class RunningController {
 	
 	// stop 눌렀을 때 redis에 있던 record를 꺼내서 db에 저장한다.
 	@PostMapping
-	public ResponseEntity saveStopRecord(@RequestBody Map<String, Object> runningInfo){
+	public ResponseEntity saveStopRecord(@RequestBody Map<String, Object> runningInfo, HttpServletRequest request){
 		System.out.println("gps/contoller/record");
-		User loginUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		
-		// 1. running 저장
-		Running savedRunning = recordService.saveRunning(runningInfo, loginUser.getUserId());
-		
-		// 2. redis에서 records 조회 + 삭제 + running의 마지막 값으로 마지막 record 추가
-		List<Record> records = recordTempRepository.findRecordByUserId(loginUser.getUserId());
-		recordTempRepository.deleteByUserId(loginUser.getUserId(), (int) savedRunning.getAccDistance());
-		Record lastRecord = Record.builder()
-								.userId(loginUser)
-								.accDistance(savedRunning.getAccDistance())
-								.accTime(savedRunning.getAccTime())
-								.build();
-		records.add(lastRecord);
-		
-		// 3. records 저장
-		int runningId = savedRunning.getRunningId();
-		recordService.saveAllRecord(runningId, records);
-		
-		// 4. 달린 지역 저장
-		List<Gugun> gugunList = recordService.saveAllGugun(savedRunning, (List<String>) runningInfo.get("gugun"));
-		
-		// 5. 총 running 업데이트
-		recordService.updateRunningUser(loginUser, savedRunning);
-		
-		HashMap<String, Object> map = new HashMap<String, Object>();
-		map.put("running", savedRunning);
-		map.put("records", records);
-		map.put("area", gugunList);
-		
-		System.out.println(loginUser.getUserId()+" 저장 "+runningId);
-		rankService.getRaceExp(loginUser.getUserId(), runningId);
-		challengeService.updateAccDistance(loginUser, savedRunning.getAccDistance());	// update위해서
-		
-		return new ResponseEntity<Response>(new 
-				Response(StatusCode.OK, ResponseMessage.RUNNING_INSERT_SUCCESS, map), HttpStatus.OK);
+		String token = request.getHeader("AUTH");
+		if(jwtTokenProvider.validateToken(token)) {
+			User loginUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+			
+			// 1. running 저장
+			Running savedRunning = recordService.saveRunning(runningInfo, loginUser.getUserId());
+			
+			// 2. redis에서 records 조회 + 삭제 + running의 마지막 값으로 마지막 record 추가
+			List<Record> records = recordTempRepository.findRecordByUserId(loginUser.getUserId());
+			recordTempRepository.deleteByUserId(loginUser.getUserId(), (int) savedRunning.getAccDistance());
+			Record lastRecord = Record.builder()
+					.userId(loginUser)
+					.accDistance(savedRunning.getAccDistance())
+					.accTime(savedRunning.getAccTime())
+					.build();
+			records.add(lastRecord);
+			
+			// 3. records 저장
+			int runningId = savedRunning.getRunningId();
+			recordService.saveAllRecord(runningId, records);
+			
+			// 4. 달린 지역 저장
+			List<Gugun> gugunList = recordService.saveAllGugun(savedRunning, (List<String>) runningInfo.get("gugun"));
+			
+			// 5. 총 running 업데이트
+			recordService.updateRunningUser(loginUser, savedRunning);
+			
+			HashMap<String, Object> map = new HashMap<String, Object>();
+			map.put("running", savedRunning);
+			map.put("records", records);
+			map.put("area", gugunList);
+			
+			System.out.println(loginUser.getUserId()+" 저장 "+runningId);
+			rankService.getRaceExp(loginUser.getUserId(), runningId);
+			challengeService.updateAccDistance(loginUser, savedRunning.getAccDistance());	// update위해서
+			
+			return new ResponseEntity<Response>(new 
+					Response(StatusCode.OK, ResponseMessage.RUNNING_INSERT_SUCCESS, map), HttpStatus.OK);
+			
+		}else {
+			return new ResponseEntity<Response>(new Response(StatusCode.FORBIDDEN, ResponseMessage.UNAUTHORIZED),
+					HttpStatus.FORBIDDEN);
+		}
 	}
 	
 	// 해당 running에 대한 기록을 보고싶을 때 running기록과 records를 쭉 보내준다.
@@ -183,18 +197,21 @@ public class RunningController {
 		int userId = 0;
 		if(jwtTokenProvider.validateToken(token)) {
 			userId = jwtTokenProvider.getUserIdFromJwt(token);
-		}
-		
-		List<User> friends = friendService.list(userId);
+			List<User> friends = friendService.list(userId);
 //		List<RunningUser> runningUsers = recordService.findAllRunningUserByUserId(friends);
-		List<Running> runnings = recordService.findRunningByFriendsId(friends);
-		
-		HashMap<String, Object> map = new HashMap<String, Object>();
-		map.put("friends", friends);
-		map.put("runnings", runnings);
-		
-		return new ResponseEntity<Response>(
-				new Response(StatusCode.OK, ResponseMessage.RUNNING_FRIENDS_RECORD, map), HttpStatus.OK);
+			List<Running> runnings = recordService.findRunningByFriendsId(friends);
+			
+			HashMap<String, Object> map = new HashMap<String, Object>();
+			map.put("friends", friends);
+			map.put("runnings", runnings);
+			
+			return new ResponseEntity<Response>(
+					new Response(StatusCode.OK, ResponseMessage.RUNNING_FRIENDS_RECORD, map), HttpStatus.OK);
+
+		}else {
+			return new ResponseEntity<Response>(new Response(StatusCode.FORBIDDEN, ResponseMessage.UNAUTHORIZED),
+					HttpStatus.FORBIDDEN);
+		}
 	}
 	
 	// running을 삭제한다.
@@ -204,12 +221,15 @@ public class RunningController {
 		int userId = 0;
 		if(jwtTokenProvider.validateToken(token)) {
 			userId = jwtTokenProvider.getUserIdFromJwt(token);
+			Long ret = recordService.deleteRunningByUserId(userId, runningId);
+			
+			return new ResponseEntity<Response>(
+					new Response(StatusCode.OK, ResponseMessage.RUNNING_DELETE_RECORD, ret), HttpStatus.FORBIDDEN);
+		
+		}else {
+			return new ResponseEntity<Response>(new Response(StatusCode.FORBIDDEN, ResponseMessage.UNAUTHORIZED),
+					HttpStatus.FORBIDDEN);
 		}
-		
-		Long ret = recordService.deleteRunningByUserId(userId, runningId);
-		
-		return new ResponseEntity<Response>(
-				new Response(StatusCode.OK, ResponseMessage.RUNNING_DELETE_RECORD, ret), HttpStatus.FORBIDDEN);
 	}
 	
 	@ApiOperation(value = "유저가 활동 지역으로 설정한 곳에서의 런닝 기록 조회", response = ResponseEntity.class)
@@ -219,12 +239,15 @@ public class RunningController {
 		int userId = 0;
 		if(jwtTokenProvider.validateToken(token)) {
 			userId = jwtTokenProvider.getUserIdFromJwt(token);
+			List<Running> runnings = recordService.findAllRunningByActivityArea(userId);
+			
+			return new ResponseEntity<Response>(new 
+					Response(StatusCode.OK, ResponseMessage.AREA_RUNNINGS_SUCCESS, runnings), HttpStatus.OK);
+		}else {
+			return new ResponseEntity<Response>(new Response(StatusCode.FORBIDDEN, ResponseMessage.UNAUTHORIZED),
+					HttpStatus.FORBIDDEN);
 		}
 		
-		List<Running> runnings = recordService.findAllRunningByActivityArea(userId);
-		
-		return new ResponseEntity<Response>(new 
-				Response(StatusCode.OK, ResponseMessage.AREA_RUNNINGS_SUCCESS, runnings), HttpStatus.OK);
 	}
 	
 	@ApiOperation(value = "유저의 총 뛴 거리, 횟수 등의 요약 정보", response = ResponseEntity.class)
@@ -234,12 +257,14 @@ public class RunningController {
 		int userId = 0;
 		if(jwtTokenProvider.validateToken(token)) {
 			userId = jwtTokenProvider.getUserIdFromJwt(token);
+			RunningUser runningUser = recordService.findRunningUserByUserId(userId);
+			
+			return new ResponseEntity<Response>(new 
+					Response(StatusCode.OK, ResponseMessage.USER_SUMMARY_RUNNING_SUCCESS, runningUser), HttpStatus.OK);
+		}else {
+			return new ResponseEntity<Response>(new Response(StatusCode.FORBIDDEN, ResponseMessage.UNAUTHORIZED),
+					HttpStatus.FORBIDDEN);
 		}
-		
-		RunningUser runningUser = recordService.findRunningUserByUserId(userId);
-		
-		return new ResponseEntity<Response>(new 
-				Response(StatusCode.OK, ResponseMessage.USER_SUMMARY_RUNNING_SUCCESS, runningUser), HttpStatus.OK);
 	}
 	
 	@ApiOperation(value = "유저의 총 뛴 거리, 횟수 등의 요약 정보", response = ResponseEntity.class)
@@ -250,7 +275,7 @@ public class RunningController {
 			int userId = jwtTokenProvider.getUserIdFromJwt(token);
 			List<RunningUser> runningUsers = recordService.findRunningUserByUserIdAndUserId(userId);
 			return new ResponseEntity<Response>(new 
-					Response(StatusCode.OK, ResponseMessage.USER_SUMMARY_RUNNING_SUCCESS, runningUsers), HttpStatus.OK);
+					Response(StatusCode.OK, ResponseMessage.REGION_SUMMARY_RUNNING_SUCCESS, runningUsers), HttpStatus.OK);
 		}else {
 			return new ResponseEntity<Response>(new 
 					Response(StatusCode.FORBIDDEN, ResponseMessage.FORBIDDEN), HttpStatus.FORBIDDEN);
